+++ src/components/index.tsx
@@ -1,18 +1,18 @@
 // components
 import Badge from '@/components/common/Badge'
 import Button from '@/components/common/Button'
-import ImageCard from '@/components/common/card/ImageCard'
+import ImageCard from '@/components/common/cardTemp/ImageCard'
 import Header from '@/components/header/Header'
 import { Symbol } from '@/components/common/Symbol'
 import { Logo } from '@/components/common/Logo'
 import { Footer } from '@/components/footer/Footer'
-import ToastContainer from '@/components/common/toast/ToastContainer'
+import ToastContainer from '@/components/common/toastTemp/ToastContainer'
 import RootLayout from '@/components/layout/RootLayout'
 import MyPageLayout from '@/components/layout/MyPageLayout'
 import Chat from '@/components/chat/Chat'
-import ListItemSkeleton from '@/components/common/skeleton/ListItemSkeleton'
-import ImageCardSkeleton from '@/components/common/skeleton/ImageCardSkeleton'
-import AvatarSkeleton from '@/components/common/skeleton/AvatarSkeleton'
+import ListItemSkeleton from '@/components/common/skeletonTemp/ListItemSkeleton'
+import ImageCardSkeleton from '@/components/common/skeletonTemp/ImageCardSkeleton'
+import AvatarSkeleton from '@/components/common/skeletonTemp/AvatarSkeleton'
 import Avatar from '@/components/common/Avatar'
 import AuthLayout from '@/components/layout/AuthLayout'
 import Notification from '@/components/notification/Notification'
